Refactor views into two files since it was getting bulky
