+++ persimmon/website/tests/common_test_functions.py
@@ -1,10 +1,11 @@
 from django.urls import reverse
-from .. import views
+
+import persimmon.website.views.apis
 
 
 def view_pending_account(client):
     req = client.post(
-            reverse(views.get_pending_bank_accounts),
+            reverse(persimmon.website.views.apis.get_pending_bank_accounts),
             content_type='application/json',
             data={})
     return req
@@ -12,7 +13,7 @@ def view_pending_account(client):
 
 def approve_account(client, req_data, number):
     req = client.post(
-        reverse(views.approve_bank_account),
+        reverse(persimmon.website.views.apis.approve_bank_account),
         content_type='application/json',
         data={'account_number': req_data[number]['account'],
               'approved': True})
