+++ persimmon/website/tests/test_scheduled.py
@@ -1,7 +1,8 @@
 from django.test import TestCase, Client
 from django.urls import reverse
+
+import persimmon.website.views.apis
 from ..models import EmployeeLevel, Appointment
-from .. import views
 from ..common import make_user
 
 class TestSchedule(TestCase):
@@ -25,13 +26,13 @@ class TestSchedule(TestCase):
 
         # creating appointments
         req = client_user1.post(
-            reverse(views.schedule),
+            reverse(persimmon.website.views.apis.schedule),
             content_type='application/json',
             data={"time": "2021-03-15 14:30"})
         self.assertEqual(req.status_code, 200)
 
         req = client_user2.post(
-            reverse(views.schedule),
+            reverse(persimmon.website.views.apis.schedule),
             content_type='application/json',
             data={"time": "2021-03-15 14:30"})
         self.assertEqual(req.status_code, 200)
@@ -56,13 +57,13 @@ class TestSchedule(TestCase):
 
         # creating appointments
         req = client_user1.post(
-            reverse(views.schedule),
+            reverse(persimmon.website.views.apis.schedule),
             content_type='application/json',
             data={"time": "2021-03-15 14:30"})
         self.assertEqual(req.status_code, 200)
 
         req = client_user2.post(
-            reverse(views.schedule),
+            reverse(persimmon.website.views.apis.schedule),
             content_type='application/json',
             data={"time": "2021-03-16 14:30"})
         self.assertEqual(req.status_code, 200)
@@ -87,13 +88,13 @@ class TestSchedule(TestCase):
 
         # creating appointments
         req = client_user1.post(
-            reverse(views.schedule),
+            reverse(persimmon.website.views.apis.schedule),
             content_type='application/json',
             data={"time": "2021-03-15 14:30"})
         self.assertEqual(req.status_code, 200)
 
         req = client_user2.post(
-            reverse(views.schedule),
+            reverse(persimmon.website.views.apis.schedule),
             content_type='application/json',
             data={"time": "2021-03-15 14:30"})
         self.assertEqual(req.status_code, 200)
