+++ persimmon/website/tests/test_session.py
@@ -1,14 +1,15 @@
 from django.test import TestCase, Client
 from django.urls import reverse
 
+import persimmon.website.views.apis
 from ..common import make_user
-from .. import views
+
 
 class TestSessions(TestCase):
     @staticmethod
     def is_logged_in(client):
         return client.post(
-            reverse(views.login_status),
+            reverse(persimmon.website.views.apis.login_status),
             content_type='application/json',
             data={}).json()['logged_in']
 
@@ -18,14 +19,14 @@ class TestSessions(TestCase):
 
         # login with invalid creds should fail
         req = client.post(
-            reverse(views.persimmon_login),
+            reverse(persimmon.website.views.apis.persimmon_login),
             content_type='application/json',
             data={"username": "username", "password": "wrong"})
         self.assertEqual(req.status_code, 200)
         req_data = req.json()
         self.assertIn("error", req_data)
         req = client.post(
-            reverse(views.persimmon_login),
+            reverse(persimmon.website.views.apis.persimmon_login),
             content_type='application/json',
             data={"username": "wrong", "password": "password"})
         self.assertEqual(req.status_code, 200)
@@ -35,7 +36,7 @@ class TestSessions(TestCase):
 
         # successful login should work
         req = client.post(
-            reverse(views.persimmon_login),
+            reverse(persimmon.website.views.apis.persimmon_login),
             content_type='application/json',
             data={"username": "username", "password": "password"})
         self.assertEqual(req.status_code, 200)
@@ -45,14 +46,14 @@ class TestSessions(TestCase):
 
         # login should not work once you're logged in
         req = client.post(
-            reverse(views.persimmon_login),
+            reverse(persimmon.website.views.apis.persimmon_login),
             content_type='application/json',
             data={"username": "username", "password": "password"})
         self.assertEqual(req.status_code, 404)
 
         # logout should work
         req = client.post(
-            reverse(views.persimmon_logout),
+            reverse(persimmon.website.views.apis.persimmon_logout),
             content_type='application/json',
             data={})
         self.assertEqual(req.status_code, 200)
@@ -60,7 +61,7 @@ class TestSessions(TestCase):
 
         # logout should not work once you're logged out
         req = client.post(
-            reverse(views.persimmon_logout),
+            reverse(persimmon.website.views.apis.persimmon_logout),
             content_type='application/json',
             data={})
         self.assertEqual(req.status_code, 404)
