+++ persimmon/website/tests/test_user_account.py
@@ -1,8 +1,8 @@
 from django.test import TestCase, Client
 from django.urls import reverse
 
+import persimmon.website.views.apis
 from ..models import User
-from .. import views
 
 
 class TestAccountWorkflow(TestCase):
@@ -15,7 +15,7 @@ class TestAccountWorkflow(TestCase):
 
         client_user1 = Client()
 
-        req1 = client_user1.post(reverse(views.create_user_account), content_type='application/json',
+        req1 = client_user1.post(reverse(persimmon.website.views.apis.create_user_account), content_type='application/json',
                                  data={"username": "gdeshpande", "first_name": "Gaurav", "last_name": "Deshpande",
                                        "password": "password", "email": "test_email", "phone": "4803333141",
                                        "address": "address"})
@@ -25,7 +25,7 @@ class TestAccountWorkflow(TestCase):
 
         client_user2 = Client()
 
-        req2 = client_user2.post(reverse(views.create_user_account), content_type='application/json',
+        req2 = client_user2.post(reverse(persimmon.website.views.apis.create_user_account), content_type='application/json',
                                  data={"username": "new_user", "last_name": "Deshpande",
                                        "password": "password", "email": "test_email", "phone": "4803333141",
                                        "address": "address"})
