+++ persimmon/website/tests/test_user_info.py
@@ -1,8 +1,8 @@
 from django.test import TestCase, Client
 from django.urls import reverse
 
+import persimmon.website.views.apis
 from ..models import EmployeeLevel
-from .. import views
 from ..common import make_user
 
 
@@ -18,7 +18,7 @@ class TestUserInformation(TestCase):
 
         # test if client can change there address successfully
         req = client.post(
-            reverse(views.change_my_address),
+            reverse(persimmon.website.views.apis.change_my_address),
             content_type='application/json',
             data={'new_address': '123 fake street'})
         self.assertEqual(req.status_code, 200)
@@ -28,7 +28,7 @@ class TestUserInformation(TestCase):
 
         # test if client can change there phone number
         req = client.post(
-            reverse(views.change_my_phone),
+            reverse(persimmon.website.views.apis.change_my_phone),
             content_type='application/json',
             data={'new_phone': '0987654321'})
         self.assertEqual(req.status_code, 200)
@@ -38,7 +38,7 @@ class TestUserInformation(TestCase):
 
         # test if client can change there email
         req = client.post(
-            reverse(views.change_my_email),
+            reverse(persimmon.website.views.apis.change_my_email),
             content_type='application/json',
             data={'new_email': 'new@example.com'})
         self.assertEqual(req.status_code, 200)
