+++ persimmon/website/views/__init__.py
@@ -1,0 +1,17 @@
+from django.http import Http404
+
+from persimmon.website.models import EmployeeLevel, User
+
+
+def current_user(request, required_auth=EmployeeLevel.CUSTOMER, expect_not_logged_in=False):
+    if expect_not_logged_in:
+        if not request.user.is_authenticated:
+            return None
+        raise Http404("API unavailable to current authentication")
+    if not request.user.is_authenticated:
+        raise Http404("API unavailable to current authentication")
+
+    user = User.objects.get(django_user=request.user)
+    if not user.check_level(required_auth):
+        raise Http404("API unavailable to current authentication")
+    return user
